delete group

+++ api/admin/group/delete.php
@@ -11,7 +11,28 @@ if ($_SERVER["REQUEST_METHOD"] == "DELETE") {
 
         if ($user) {
             $data = json_decode(file_get_contents("php://input", true));
+            
+            if(!$data || !$data->id){
+                echo json_encode([
+                    'code' => 400,
+                    'message' => 'Vui lòng cung cấp id của nhóm câu hỏi!'                    
+                ]);
+                return;
+            }
+            
             $id = $data->id;
+            $obj->select("groups","*",null,"id={$id}",null,null,null);
+            $result = $obj->getResult();
+            if(!$result){
+                echo json_encode([
+                    'code' => 404,
+                    'message' => 'Không tìm thấy nhóm câu hỏi phù hợp!'                    
+                ]);
+                return;
+            }
+
+
+           
            
             $obj->delete('groups',"id={$id}");
